refactor(web): share the branch-detection test router and tighten comments
Point the Go contract's hand-mirror comments at web/src/util/commit.ts,
where FEEDBACK_OPEN_COMMIT_MESSAGE and shimUpdateCommitMessage now live —
those comments are the only cross-language link, so they move with the
code.

Extract the branch-mode read router both hook tests hand-rolled into
@/test/branchDetectionClient. The student-hook copy resolved null for a
not-accepted repo, which bypassed getRepo's tolerateGitHubError; the
shared one rejects 404 like the real client, so both tests now exercise
that path.

Type shimUpdateCommitMessage with SubmissionMode instead of respelling
the union, read commit subjects with indexOf/slice rather than splitting
every line to keep one, and derive the timeline subject once instead of
twice per commit on a per-keystroke path.
