+++ web/src/hooks/useDetectedSubmissions.test.tsx
@@ -14,25 +14,7 @@ import {
   FEEDBACK_OPEN_COMMIT_MESSAGE,
   shimUpdateCommitMessage,
 } from "@/util/commit"
-import { GitHubAPIError, type GitHubRateLimit } from "@/github-core/errors"
-
-const noRateLimit: GitHubRateLimit = {
-  limit: null,
-  remaining: null,
-  used: null,
-  reset: null,
-  resource: null,
-  retryAfter: null,
-}
-
-const apiError = (status: number) =>
-  new GitHubAPIError({
-    status,
-    url: "https://api.github.com/x",
-    message: `HTTP ${status}`,
-    body: null,
-    rateLimit: noRateLimit,
-  })
+import { apiError, branchClient } from "@/test/branchDetectionClient"
 
 const makeClient = () =>
   new QueryClient({ defaultOptions: { queries: { retry: false } } })
@@ -48,39 +30,6 @@ const base = {
   classroom: "cs101",
   assignment: "hw1",
   enabled: true,
-}
-
-// A URL router for the branch-mode reads: repo object (default_branch),
-// the marker commit list (baseline), and the default-branch commit log.
-// Branch commits get a representative `commit` payload (the real list-commits
-// response always carries one) so detection can read the commit time.
-function branchClient(opts: {
-  defaultBranch?: string | null
-  baselineCommits?: Array<{ sha: string }>
-  branchCommits?: Array<{ sha: string; message?: string }>
-}) {
-  return (url: string) => {
-    if (/\/repos\/[^/]+\/[^/]+$/.test(url)) {
-      return opts.defaultBranch === null
-        ? Promise.reject(apiError(404))
-        : Promise.resolve({ default_branch: opts.defaultBranch ?? "main" })
-    }
-    if (url.includes("path=.classroom50.yaml")) {
-      return Promise.resolve(opts.baselineCommits ?? [])
-    }
-    if (url.includes("/commits?sha=")) {
-      return Promise.resolve(
-        (opts.branchCommits ?? []).map((c) => ({
-          sha: c.sha,
-          commit: {
-            message: c.message ?? c.sha,
-            committer: { date: "2026-06-20T10:00:00Z" },
-          },
-        })),
-      )
-    }
-    return Promise.resolve([])
-  }
 }
 
 beforeEach(() => {
