+++ web/src/pages/assignments/sections/SubmissionsSubsection.tsx
@@ -57,9 +57,9 @@ export function SubmissionsSubsection({
   )
 }
 
-// The submission definition: branch (every default-branch commit counts, the
-// shim grades every push) vs tag (on submit only). Always visible. The retrofit edit warning only shows when a
-// built-in shim exists (there is a shim to update).
+// The submission definition: branch (every default-branch student commit
+// counts) vs tag (on submit only). Always visible. The retrofit edit warning
+// only shows when a built-in shim exists (there is a shim to update).
 function SubmissionModeField({
   form,
   edit,
